nudgeme-server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
apscheduler>=3.10.4
pydantic>=2.10.0
gunicorn==22.0.0

    